server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
httpx==0.27.0
supabase==2.7.0
pydantic==2.9.0
python-dotenv==1.0.0
pymysql==1.1.0
redis==5.0.0
oss2==2.18.0
cryptography==41.0.0
